# Add LRU cache to `is_valid_license_info`