--- a/openverse_catalog/dags/common/licenses/licenses.py
+++ b/openverse_catalog/dags/common/licenses/licenses.py
@@ -226,6 +226,7 @@ def _build_license_url(license_path) -> str:
     return rewritten_license_url
 
 
+@lru_cache(maxsize=1024)
 def is_valid_license_info(license_info: LicenseInfo) -> bool:
     base_path = "https://creativecommons.org/"
     try: