@AetherUnbound 2022-03-21
> Nice! I was expecting a larger effect on performance though, I wonder why it's only two-fold? thinking

I think it's because at the end of the day, this is just a dictionary lookup. So it's already pretty fast to begin with!